Add useUserCred hook and use userCred to check for safe mode in admin panels
