+++ tsconfig.json
@@ -22,6 +22,9 @@
       "@styles/*": [
         "styles/*"
       ],
+      "@handlers/*": [
+        "handlers/*"
+      ],
       "@components/*": [
         "components/*"
       ],
